\name{batade-package}
\alias{batade-package}
\alias{batade}
\docType{package}
\title{
HTML reports and so on
}
\description{
This package provides some utility functions (e.g HTML report maker).
}
\details{
\tabular{ll}{
Package: \tab batade\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-06-13\cr
License: \tab GPL-2\cr
}
}
\author{
Ichikawa Daisuke

Maintainer: Ichikawa Daisuke <ichikawadaisuke at gmail.com>

}